Implement a two-pass parse
First pass uses the faster SLL prediction, the second pass falls back to LL for more complex expressions at the cost of performance.

It should be noted this will be slower when there the number of expected parsing failure is higher than expected successes.  It seems like we should optimize for success to ensure overall evaluation is faster.

see: https://tomassetti.me/improving-the-performance-of-an-antlr-parser/